+++ src/array/codec/bytes_to_bytes/zstd.rs
@@ -39,6 +39,10 @@ mod tests {
         let encoded = codec.encode(bytes.clone()).unwrap();
         let decoded = codec.decode(encoded, &bytes_representation).unwrap();
         assert_eq!(bytes, decoded);
+
+        let encoded = codec.par_encode(bytes.clone()).unwrap();
+        let decoded = codec.par_decode(encoded, &bytes_representation).unwrap();
+        assert_eq!(bytes, decoded);
     }
 
     #[test]